requirements.txt
SQLAlchemy==1.0.11
mysql-connector-python-rf==2.1.3
numpy==1.10.2
pandas==0.17.1
python-dateutil==2.4.2
pytz==2015.7
requests==2.9.1
six==1.10.0
wsgiref==0.1.2
